move Mockutils out of unit-tests since it is used in a few integration tests
