+++ test/suite/commands/NewFile.test.ts
@@ -17,7 +17,7 @@ import { expect } from "chai";
 import { match } from "sinon";
 import * as path from "path";
 import { newSwiftFile } from "../../../src/commands/newFile";
-import { mockGlobalObject } from "../../unit-tests/MockUtils";
+import { mockGlobalObject } from "../../MockUtils";
 import { TemporaryFolder } from "../../../src/utilities/tempFolder";
 import { fileExists } from "../../../src/utilities/filesystem";
