+++ test/suite/testexplorer/TestExplorerIntegration.test.ts
@@ -30,7 +30,7 @@ import { WorkspaceContext } from "../../../src/WorkspaceContext";
 import { TestRunProxy } from "../../../src/TestExplorer/TestRunner";
 import { Version } from "../../../src/utilities/version";
 import { TestKind } from "../../../src/TestExplorer/TestKind";
-import { mockGlobalObject } from "../../unit-tests/MockUtils";
+import { mockGlobalObject } from "../../MockUtils";
 import {
     MessageRenderer,
     TestSymbol,
