+++ test/unit-tests/ui/SwiftBuildStatus.test.ts
@@ -22,7 +22,7 @@ import {
     mockGlobalEvent,
     instance,
     mockFn,
-} from "../MockUtils";
+} from "../../MockUtils";
 import { SwiftExecution } from "../../../src/tasks/SwiftExecution";
 import { TestSwiftProcess } from "../../fixtures";
 import { StatusItem } from "../../../src/ui/StatusItem";
